refactor: fix tests 2
