+++ packages/utils/src/lib/report-to-stdout.ts
@@ -2,13 +2,13 @@ import chalk from 'chalk';
 import cliui from 'cliui';
 import { Report } from '@code-pushup/models';
 import { NEW_LINE } from './md';
-import { CODE_PUSHUP_DOMAIN, FOOTER_PREFIX } from './report';
 import {
+  CODE_PUSHUP_DOMAIN,
+  FOOTER_PREFIX,
   countWeightedRefs,
-  reportHeadlineText,
-  reportOverviewTableHeaders,
   sumRefs,
-} from './utils';
+} from './report';
+import { reportHeadlineText, reportOverviewTableHeaders } from './utils';
 
 const ui = cliui({ width: 60 }); // @TODO check display width
