+++ packages/utils/src/lib/report.ts
@@ -1,2 +1,58 @@
+import { CategoryConfig, Issue } from '@code-pushup/models';
+import { pluralize } from './utils';
+
 export const FOOTER_PREFIX = 'Made with ❤️ by';
 export const CODE_PUSHUP_DOMAIN = 'code-pushup.dev';
+
+export function slugify(text: string): string {
+  return text
+    .trim()
+    .toLowerCase()
+    .replace(/\s+|\//g, '-')
+    .replace(/[^a-z0-9-]/g, '');
+}
+
+export function formatBytes(bytes: number, decimals = 2) {
+  if (!+bytes) return '0 B';
+
+  const k = 1024;
+  const dm = decimals < 0 ? 0 : decimals;
+  const sizes = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
+}
+
+export function calcDuration(start: number, stop?: number): number {
+  stop = stop !== undefined ? stop : performance.now();
+  return Math.floor(stop - start);
+}
+
+export function formatCount(count: number, name: string) {
+  const text = count === 1 ? name : pluralize(name);
+  return `${count} ${text}`;
+}
+
+export function countWeightedRefs(refs: CategoryConfig['refs']) {
+  return refs
+    .filter(({ weight }) => weight > 0)
+    .reduce((sum, { weight }) => sum + weight, 0);
+}
+
+export function compareIssueSeverity(
+  severity1: Issue['severity'],
+  severity2: Issue['severity'],
+): number {
+  const levels: Record<Issue['severity'], number> = {
+    info: 0,
+    warning: 1,
+    error: 2,
+  };
+  return levels[severity1] - levels[severity2];
+}
+
+// @TODO replace with real scoring logic
+export function sumRefs(refs: CategoryConfig['refs']) {
+  return refs.reduce((sum, { weight }) => sum + weight, 0);
+}
