+++ packages/utils/src/lib/utils.ts
@@ -1,51 +1,22 @@
 import { readFile } from 'fs/promises';
-import { CategoryConfig, Issue } from '@code-pushup/models';
 
 export const reportHeadlineText = 'Code Pushup Report';
 export const reportOverviewTableHeaders = ['Category', 'Score', 'Audits'];
 
-// @TODO replace with real scoring logic
-export function sumRefs(refs: CategoryConfig['refs']) {
-  return refs.reduce((sum, { weight }) => sum + weight, 0);
-}
+// === Transform
 
-export function countWeightedRefs(refs: CategoryConfig['refs']) {
-  return refs
-    .filter(({ weight }) => weight > 0)
-    .reduce((sum, { weight }) => sum + weight, 0);
-}
-
-export function formatBytes(bytes: number, decimals = 2) {
-  if (!+bytes) return '0 B';
-
-  const k = 1024;
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
-}
-
-export function calcDuration(start: number, stop?: number): number {
-  stop = stop !== undefined ? stop : performance.now();
-  return Math.floor(stop - start);
-}
-
-export function distinct<T extends string | number | boolean>(array: T[]): T[] {
-  return Array.from(new Set(array));
+export function pluralize(text: string): string {
+  if (text.endsWith('y')) {
+    return text.slice(0, -1) + 'ies';
+  }
+  if (text.endsWith('s')) {
+    return `${text}es`;
+  }
+  return `${text}s`;
 }
 
 export function toArray<T>(val: T | T[]): T[] {
   return Array.isArray(val) ? val : [val];
-}
-
-export function slugify(text: string): string {
-  return text
-    .trim()
-    .toLowerCase()
-    .replace(/\s+|\//g, '-')
-    .replace(/[^a-z0-9-]/g, '');
 }
 
 export function objectToEntries<T extends object>(obj: T) {
@@ -61,17 +32,13 @@ export function countOccurrences<T extends PropertyKey>(
   );
 }
 
-export function compareIssueSeverity(
-  severity1: Issue['severity'],
-  severity2: Issue['severity'],
-): number {
-  const levels: Record<Issue['severity'], number> = {
-    info: 0,
-    warning: 1,
-    error: 2,
-  };
-  return levels[severity1] - levels[severity2];
+// === Validation
+
+export function distinct<T extends string | number | boolean>(array: T[]): T[] {
+  return Array.from(new Set(array));
 }
+
+// === File
 
 export async function readTextFile(path: string): Promise<string> {
   const buffer = await readFile(path);
@@ -81,19 +48,4 @@ export async function readTextFile(path: string): Promise<string> {
 export async function readJsonFile(path: string): Promise<unknown> {
   const text = await readTextFile(path);
   return JSON.parse(text);
-}
-
-export function formatCount(count: number, name: string) {
-  const text = count === 1 ? name : pluralize(name);
-  return `${count} ${text}`;
-}
-
-export function pluralize(text: string): string {
-  if (text.endsWith('y')) {
-    return text.slice(0, -1) + 'ies';
-  }
-  if (text.endsWith('s')) {
-    return `${text}es`;
-  }
-  return `${text}s`;
 }
